+++ ukat/vessels/phase_contrast.py
@@ -248,7 +248,7 @@ class PhaseContrast:
                     nib.save(velocity_nifti, base_path +
                              '_velocity_array.nii.gz')
                 elif result == 'mask':
-                    mask_nifti = nib.Nifti1Image(self.mask.astype(int),
+                    mask_nifti = nib.Nifti1Image(self.mask.astype(np.uint16),
                                                  affine=self.affine)
                     nib.save(mask_nifti, base_path + '_mask.nii.gz')
         else:
